Make Results component more common, specify exports in index.ts
