+++ mathesar_ui/src/systems/query-builder/types.ts
@@ -1,0 +1,2 @@
+export type { default as QueryManager } from './QueryManager';
+export type { default as QueryRunner } from './QueryRunner';
